refactor: split assistant tools into @mongodb-js/mcp-tools-assistant package
Extract AssistantToolBase and concrete tools (SearchKnowledgeTool,
ListKnowledgeSourcesTool) from src/tools/assistant/ into an isolated
workspace package at packages/tools-assistant/.

Follows the pattern established by mcp-tools-atlas-local:
- Package exports AssistantTools array for root registration
- Domain-specific base class (AssistantToolBase) with config interface
- Tools grouped by operation type (read/)
- Co-located barrel exports (index.ts, tools/tools.ts)
- Unit tests remain at root for integration testing
- Accuracy tests updated to import from new package

Root package updated to depend on @mongodb-js/mcp-tools-assistant and
src/tools/index.ts refactored to import from the new package instead of
local directory.

This decouples assistant-specific code from the main server and enables
future splitting of other tool groups (atlas, mongodb) following the same
pattern.
