+++ packages/tools-assistant/package.json
@@ -1,0 +1,34 @@
+{
+    "name": "@mongodb-js/mcp-tools-assistant",
+    "version": "1.11.0-prerelease.1",
+    "private": true,
+    "type": "module",
+    "description": "Assistant tools for the MongoDB MCP server",
+    "exports": {
+        ".": {
+            "types": "./dist/index.d.ts",
+            "default": "./dist/index.js"
+        }
+    },
+    "main": "./dist/index.js",
+    "types": "./dist/index.d.ts",
+    "files": ["dist"],
+    "scripts": {
+        "compile": "tsc --build tsconfig.json",
+        "test": "vitest --run"
+    },
+    "dependencies": {
+        "@mongodb-js/devtools-proxy-support": "^0.7.8",
+        "@mongodb-js/mcp-core": "workspace:*",
+        "@mongodb-js/mcp-logging": "workspace:*",
+        "@mongodb-js/mcp-types": "workspace:*",
+        "yaml": "^2.8.1",
+        "zod": "^4.0.0"
+    },
+    "devDependencies": {
+        "@modelcontextprotocol/sdk": "^1.29.0",
+        "@types/node": "^25.6.0",
+        "typescript": "^5.9.2",
+        "vitest": "^4.1.4"
+    }
+}
