+++ packages/tools-assistant/src/tools/read/listKnowledgeSources.ts
@@ -1,6 +1,7 @@
 import type { CallToolResult } from "@modelcontextprotocol/sdk/types.js";
-import { formatUntrustedData, type OperationType, type ToolCategory } from "../tool.js";
-import { AssistantToolBase } from "./assistantTool.js";
+import { formatUntrustedData } from "@mongodb-js/mcp-core";
+import type { OperationType, ToolCategory } from "@mongodb-js/mcp-core";
+import { AssistantToolBase } from "../../assistantTool.js";
 import { LogId } from "@mongodb-js/mcp-logging";
 import { stringify as yamlStringify } from "yaml";
