+++ packages/tools-assistant/src/tools/read/searchKnowledge.ts
@@ -1,7 +1,8 @@
 import { z } from "zod";
-import { type CallToolResult } from "@modelcontextprotocol/sdk/types.js";
-import { type ToolArgs, type OperationType, type ToolCategory, formatUntrustedData } from "../tool.js";
-import { AssistantToolBase } from "./assistantTool.js";
+import type { CallToolResult } from "@modelcontextprotocol/sdk/types.js";
+import { formatUntrustedData } from "@mongodb-js/mcp-core";
+import type { ToolArgs, OperationType, ToolCategory } from "@mongodb-js/mcp-core";
+import { AssistantToolBase } from "../../assistantTool.js";
 import { LogId } from "@mongodb-js/mcp-logging";
 import { stringify as yamlStringify } from "yaml";
