+++ packages/tools-assistant/src/tools/tools.ts
@@ -1,0 +1,2 @@
+export { SearchKnowledgeTool } from "./read/searchKnowledge.js";
+export { ListKnowledgeSourcesTool } from "./read/listKnowledgeSources.js";
