+++ packages/tools-assistant/tsconfig.json
@@ -1,0 +1,17 @@
+{
+    "extends": "../../tsconfig.build.json",
+    "compilerOptions": {
+        "rootDir": "./src",
+        "outDir": "./dist",
+        "baseUrl": ".",
+        "erasableSyntaxOnly": true,
+        "composite": true
+    },
+    "include": ["src/**/*.ts"],
+    "exclude": ["dist", "node_modules", "src/**/*.test.ts"],
+    "references": [
+        { "path": "../types" },
+        { "path": "../core" },
+        { "path": "../logging" }
+    ]
+}
