--- src/tools/assistant/tools.ts
@@ -1,4 +1,0 @@
-import { ListKnowledgeSourcesTool } from "./listKnowledgeSources.js";
-import { SearchKnowledgeTool } from "./searchKnowledge.js";
-
-export { ListKnowledgeSourcesTool, SearchKnowledgeTool };
