+++ src/tools/index.ts
@@ -1,14 +1,15 @@
 import * as AtlasTools from "./atlas/tools.js";
 import { AtlasLocalTools } from "@mongodb-js/mcp-tools-atlas-local";
+import { AssistantTools } from "@mongodb-js/mcp-tools-assistant";
 import * as MongoDbTools from "./mongodb/tools.js";
-import * as AssistantTools from "./assistant/tools.js";
 import type { ToolClass } from "./tool.js";
 
 // Export the collection of tools for easier reference
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const AllTools: ToolClass<any, any, any>[] = [
-    ...Object.values({ ...MongoDbTools, ...AtlasTools, ...AssistantTools }),
+    ...Object.values({ ...MongoDbTools, ...AtlasTools }),
     ...AtlasLocalTools,
+    ...AssistantTools,
 ];
 
 export { MongoDBToolBase } from "./mongodb/mongodbTool.js";
@@ -16,8 +17,8 @@ export { MongoDBToolBase } from "./mongodb/mongodbTool.js";
 // Export all the individual tools for handpicking
 export * from "./atlas/tools.js";
 export * from "@mongodb-js/mcp-tools-atlas-local";
+export * from "@mongodb-js/mcp-tools-assistant";
 export * from "./mongodb/tools.js";
-export * from "./assistant/tools.js";
 
 // Export the base tool class and supporting types.
 export {
